Download decrypted file using StreamSaver.js
Streaming file downloads requires a server in order to instantiate a
service worker. This unfortunately means that the webapp cannot be run
statically without a network connection.

At some point, it would be useful to offer an alternative where the
entire file is decrypted into memory and then downloaded. This would
need to be gated on file size.