Pin age revision that fixes the crash
The problem was that the Decryptor type under the hood was storing a
struct with two fields:

{
    inner: R,
    header: Header,
}

The type R is instantiated as shim::StreamReader<'a> in this webapp,
which is a wrapper around wasm_streams::readable::IntoStream<'a>. My
suspicion is that the lifetime is not being handled correctly when
passing the object back to JavaScript, and the next time that we enter
the Rust code, the contents of `inner` is undefined, potentially causing
header to be unreadable.

The fixed revision of age swaps the order of `inner` and `header` in the
Decryptor struct. This doesn't actually fix the underlying issue, or
explain why an unrelated refactor triggered this issue in the first
place, but it's sufficient to allow `header` to be read correctly, and
also indicates that the method being used to store the Decryptor in
JavaScript is working.