Correct the age::Decryptor concrete type
A refactor in the upstream age branch changed how armoring was included,
and removed the need for a concrete BufReader type. This was not
reflected in the JavaScript newtype, causing the Rust structs to be
mis-interpreted in memory. We now use a type alias to ensure that the
types in Decryptor::{new, as_ref, into_box} are all consistent.