Add direct dependency on getrandom 0.1
In str4d/wage#30 we migrated our main dependencies from getrandom 0.1
to 0.2. However, x25519-dalek still depends on rand_core 0.5, which
means we still have getrandom 0.1 in our dependency tree, and need to
keep it as a direct dependency in order to enable WASM support for it.
This broke the "New Identity" button on the encrypt-to-recipients view.